requirements.txt
fastapi
uvicorn[standard]
faiss-cpu
pillow
numpy
torch
torchvision
sentence-transformers
open_clip_torch
requests
